work on leave module testing issues

+++ src/app/(app)/leaves/actions/delete-leave-request.ts
@@ -4,10 +4,14 @@ import { Tables } from "@/lib/db";
 import { Routes } from "@/lib/routes";
 import { getSupabaseClient } from "@/utils/supabase/supabaseClient";
 import { revalidatePath } from "next/cache";
+import { hasLeaveRequestPermission } from "./has-leave-request-permission";
 
 export const deleteLeaveRequest = async (requestId: number) => {
   const supabase = await getSupabaseClient();
-  const { error } = await supabase
+  const isLeaveRequestAllowed = await hasLeaveRequestPermission(requestId)
+
+  if(isLeaveRequestAllowed){
+    const { error } = await supabase
     .from(Tables.Leaves)
     .delete()
     .eq("id", requestId);
@@ -16,4 +20,6 @@ export const deleteLeaveRequest = async (requestId: number) => {
     throw error.message;
   }
   revalidatePath(Routes.LeaveRequest);
+  }
+  return;
 };
